Move sbt 1 code to main folder
